requirements.txt
hydra-core==1.3.2
ipdb==0.13.13
kornia==0.7.1
matplotlib==3.8.4
numpy==2.1.1
omegaconf==2.3.0
opencv_python==4.9.0.80
Pillow==9.5.0
pytorch_lightning==1.7.6
PyYAML==6.0.1
PyYAML==6.0.2
scikit-image==0.22.0
torch==1.12.1
torchvision==0.13.1
wandb==0.16.2
